ID-1: pages folder renamed to main_page
