+++ src/main_page/BottomSectionMainPage/BottomSectionMainPage.css
