+++ src/main_page/BottomSectionMainPage/BottomSectionMainPage.js
