+++ src/main_page/TopSectionMainPage/TopSectionMainPage.css
