+++ src/main_page/TopSectionMainPage/TopSectionMainPage.js
