+++ src/poetry/PoemPage.js
@@ -2,7 +2,7 @@ import React from 'react'
 import StaticContentService from '../StaticContentServiceYaml'
 
 // for formatting
-import './../pages/TopSectionMainPage/TopSectionMainPage.css';
+import '../main_page/TopSectionMainPage/TopSectionMainPage.css';
 
 import Divider from '@material-ui/core/Divider';
